requirements.txt
torch
torchvision
pandas
numpy
Pillow
tkinter
